Change `require lazy-seq.init-macros` calls to in-macro import-macro calls
Calling `require` in the body of a macro has unpredictable behaviour in
some runtimes (hotpot, aniseed) and seems discoraged by the API
documentation and sandbox.

Instead we can generate the relative path as a local variable (at macro
compile time) and swap the compile-time require for a macro-call-time
`import-macros` call with the relative path injected down.